Gallery 리스트 페이징, 업로드, 삭제 최적화 페이징 필요한 리스트 전부 MethodArgumentResolver 에서 처리

+++ src/main/java/me/yh/community/entity/GalleryFile.java
@@ -33,10 +33,10 @@ public class GalleryFile {
     @Column(nullable = false)
     private int fileSize;
 
-    public GalleryFile(Gallery gallery, String fileName, String originalFileName, int fileSize) {
+    public GalleryFile(Gallery gallery, String fileName, String originalFileName, long fileSize) {
         this.gallery = gallery;
         this.fileName = fileName;
         this.originalFileName = originalFileName;
-        this.fileSize = fileSize;
+        this.fileSize = (int)fileSize;
     }
 }
